Services/SmartDataService/requirements.txt
kafka-python
sparqlwrapper
rdflib-jsonld
requests